requirements.txt
basicsr==1.4.2
filetype==1.2.0
gradio==3.50.2
numpy==1.26.2
onnx==1.15.0
onnxruntime_gpu==1.17.0
# ort-nightly-gpu==1.17.0.dev20231205004
opencv-python==4.8.1.78
psutil==5.9.6
realesrgan==0.3.0
torch==2.1.2
tqdm==4.66.1
coloredlogs==15.0.1
torch --pre --index-url https://download.pytorch.org/whl/cu121
torchvision --pre --index-url https://download.pytorch.org/whl/cu121
torchaudio --pre --index-url https://download.pytorch.org/whl/cu121
tqdm==4.66.1
ffmpeg
